Update 24_fsx_modeltrainer: PERSISTENT_2 FSx with DRA, fix OOM

+++ 24_fsx_modeltrainer/scripts/train.py
@@ -103,7 +103,7 @@ def training_function(args):
         bias=False,
     )
 
-    model = AutoModelForCausalLM.from_config(config)
+    model = AutoModelForCausalLM.from_config(config, torch_dtype=torch.bfloat16)
     tokenizer = AutoTokenizer.from_pretrained(args.model_id)
 
     train_dataset = dataset["train"]
